pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi>=0.135.3",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.12.5",
    "pymongo>=4.16.0",
    "python-jose>=3.5.0",
    "redis[hiredis]>=7.4.0",
    "uvicorn>=0.42.0",
]
